+++ hil-test/Cargo.toml
@@ -153,6 +153,8 @@ hex-literal         = "0.4.1"
 nb                  = "1.1.0"
 p192                = { version = "0.13.0", default-features = false, features = ["arithmetic"] }
 p256                = { version = "0.13.2", default-features = false, features = ["arithmetic"] }
+sha1                = { version = "0.10.6", default-features = false }
+sha2                = { version = "0.10.8", default-features = false }
 
 [features]
 default = ["async", "embassy"]
